Test booting ISO as USB stick
When openQA connects the iso files, it emulates the CD-ROM drive.
In the past, issues were reported that an ISO burnt to a flash drive
would not boot properly. Forcing openQA to emulate a USB flash drive
when booting and installing a system could help us catch these issue
in the future.

However, there is a small problem here. When booting from an emulated
optical drive or USB stick, openQA gives it boot priority throughout
the life of the VM. For optical install tests this isn't a problem
because anaconda automatically ejects the 'disc' at the end of the
install process, but there's no equivalent for USB, so if we just
reboot after install, we boot back to the installer.

To resolve this, after install, we go to a console, cleanly stop the
system with `systemctl halt`, then disconnect the USB stick from the
VM using a qmp command that is now exposed by os-autoinst.

Fixes: https://pagure.io/fedora-qa/os-autoinst-distri-fedora/issue/34

+++ tests/_do_install_and_reboot.pm
@@ -150,6 +150,7 @@ sub run {
     push(@actions, 'consoletty0') if (get_var("ARCH") eq "aarch64");
     push(@actions, 'abrt') if (get_var("ABRT", '') eq "system");
     push(@actions, 'rootpw') if (get_var("INSTALLER_NO_ROOT"));
+    push(@actions, 'usbhalt') if (get_var("USBBOOT"));
     # FIXME: remove plymouth from Server install_default_upload on
     # non-aarch64 to work around RHBZ #1933378
     unless (get_var("ARCH") eq "aarch64") {
@@ -224,6 +225,16 @@ sub run {
     }
     if (grep { $_ eq 'noplymouth' } @actions) {
         assert_script_run "chroot $mount dnf -y remove plymouth";
+    }
+    if (grep { $_ eq 'usbhalt' } @actions) {
+        # halt the system cleanly to ensure install is complete
+        type_string "systemctl halt\n";
+        wait_still_screen 5;
+        # disconnect the USB stick
+        disconnect_usb;
+        # reboot via ACPI
+        power "reset";
+        return;
     }
     type_string "reboot\n" if (grep { $_ eq 'reboot' } @actions);
 }
